Move `tickets` command to owner cog
